+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/modules/InheritanceFinderModule.groovy
@@ -2,7 +2,6 @@ package ru.vyarus.guice.persist.orient.support.modules
 
 import com.google.inject.AbstractModule
 import ru.vyarus.guice.persist.orient.FinderModule
-import ru.vyarus.guice.persist.orient.support.finder.inheritance.PowerFinder
 
 /**
  * @author Vyacheslav Rusakov 
@@ -13,6 +12,6 @@ class InheritanceFinderModule extends AbstractModule {
     @Override
     protected void configure() {
         install(new PackageSchemeModule())
-        install(new FinderModule(PowerFinder))
+        install(new FinderModule())
     }
 }