+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/modules/TestFinderModule.groovy
@@ -2,8 +2,6 @@ package ru.vyarus.guice.persist.orient.support.modules
 
 import com.google.inject.AbstractModule
 import ru.vyarus.guice.persist.orient.FinderModule
-import ru.vyarus.guice.persist.orient.support.finder.FinderWithPlaceholders
-import ru.vyarus.guice.persist.orient.support.finder.InterfaceFinder
 
 /**
  * @author Vyacheslav Rusakov 
@@ -14,6 +12,6 @@ class TestFinderModule extends AbstractModule {
     @Override
     protected void configure() {
         install(new PackageSchemeModule())
-        install(new FinderModule(InterfaceFinder, FinderWithPlaceholders))
+        install(new FinderModule())
     }
 }